+++ docs/changelog.rst
@@ -1,6 +1,6 @@
 .. _changelog:
 
-Change log
+Change Log
 ==========
 
 0.3.0 (2018-??-??)
@@ -14,6 +14,7 @@ Change log
 - ``manage initdb`` 会自动加载foo模块及子模块中所有声明的 ``db.Model``
 - Flask app的name更正为项目的名称
 - 项目配置相关文件统一命名为app，包括 conf/app.py , conf/app-env.sh , bin/app-config.sh
+- 配置项由原来的 ``config.settings`` 改为直接通过 ``settings`` 对象获取
 
 0.2.5 (2018-09-18)
 ------------------